Move ModuleInstallDescriptor tests to ModuleInstallDescriptorTests
