+++ Tests/Core/Types/ModuleInstallDescriptorTests.cs
@@ -1,8 +1,13 @@
+using System.IO;
+using System.Linq;
+
 using NUnit.Framework;
 using Newtonsoft.Json;
+using ICSharpCode.SharpZipLib.Zip;
 
 using CKAN;
 using CKAN.Games.KerbalSpaceProgram;
+using Tests.Data;
 
 namespace Tests.Core.Types
 {
@@ -51,5 +56,113 @@ namespace Tests.Core.Types
             Assert.That(act, Throws.Exception);
         }
 
+        [Test]
+        public void DefaultInstallStanza_KSP1_GameData()
+        {
+            // Arrange
+            var game = new KerbalSpaceProgram();
+            var stanza = ModuleInstallDescriptor.DefaultInstallStanza(game, "DogeCoinFlag");
+            // Same again, but screwing up the case (we see this *all the time*)
+            var stanza2 = ModuleInstallDescriptor.DefaultInstallStanza(game, "DogecoinFlag");
+
+            // Assert
+            Assert.AreEqual("GameData",     stanza.install_to);
+            Assert.AreEqual("DogeCoinFlag", stanza.find);
+            Assert.AreEqual("GameData",     stanza2.install_to);
+            Assert.AreEqual("DogecoinFlag", stanza2.find);
+        }
+
+        [TestCase("Ships")]
+        [TestCase("Ships/VAB")]
+        [TestCase("Ships/SPH")]
+        [TestCase("Ships/@thumbs")]
+        [TestCase("Ships/@thumbs/VAB")]
+        [TestCase("Ships/@thumbs/SPH")]
+        [TestCase("Ships/Script")]
+        public void AllowsInstallsToShipsDirectories(string directory)
+        {
+            // Arrange
+            var zip = ZipFile.Create(new MemoryStream());
+            zip.BeginUpdate();
+            zip.AddDirectory("ExampleShips");
+            zip.Add(new ZipEntry("ExampleShips/AwesomeShip.craft") { Size = 0, CompressedSize = 0 });
+            zip.CommitUpdate();
+
+            var mod = CkanModule.FromJson(string.Format(
+                @"{{
+                    ""spec_version"": 1,
+                    ""identifier"": ""AwesomeMod"",
+                    ""author"": ""AwesomeModder"",
+                    ""version"": ""1.0.0"",
+                    ""download"": ""https://awesomemod.example/AwesomeMod.zip"",
+                    ""install"": [
+                        {{
+                            ""file"": ""ExampleShips/AwesomeShip.craft"",
+                            ""install_to"": ""{0}""
+                        }}
+                    ]
+                }}",
+                directory));
+
+            using (var inst = new DisposableKSP())
+            {
+                // Act
+                var results = mod.install!.SelectMany(i => i.FindInstallableFiles(zip, inst.KSP))
+                                          .ToArray();
+
+                // Assert
+                CollectionAssert.AreEquivalent(
+                    new string[]
+                    {
+                        inst.KSP.ToAbsoluteGameDir($"{directory}/AwesomeShip.craft"),
+                    },
+                    results.Select(f => f.destination));
+            }
+        }
+
+        // TODO: It would be nice to merge this and the above function into one super
+        // test.
+        [Test]
+        public void AllowInstallsToScenarios()
+        {
+            // Arrange
+            // Bogus zip with example to install.
+            var zip = ZipFile.Create(new MemoryStream());
+            zip.BeginUpdate();
+            zip.AddDirectory("saves");
+            zip.AddDirectory("saves/scenarios");
+            zip.Add(new ZipEntry("saves/scenarios/AwesomeRace.sfs") { Size = 0, CompressedSize = 0 });
+            zip.CommitUpdate();
+
+            var mod = CkanModule.FromJson(@"
+                {
+                    ""spec_version"": ""v1.14"",
+                    ""identifier"": ""AwesomeMod"",
+                    ""author"": ""AwesomeModder"",
+                    ""version"": ""1.0.0"",
+                    ""download"": ""https://awesomemod.example/AwesomeMod.zip"",
+                    ""install"": [
+                        {
+                            ""file"": ""saves/scenarios/AwesomeRace.sfs"",
+                            ""install_to"": ""Scenarios""
+                        }
+                    ]
+                }");
+
+            using (var inst = new DisposableKSP())
+            {
+                // Act
+                var results = mod.install!.SelectMany(i => i.FindInstallableFiles(zip, inst.KSP))
+                                          .ToArray();
+
+                // Assert
+                CollectionAssert.AreEquivalent(
+                    new string[]
+                    {
+                        inst.KSP.ToAbsoluteGameDir("saves/scenarios/AwesomeRace.sfs"),
+                    },
+                    results.Select(f => f.destination));
+            }
+        }
     }
 }
